+++ src/common/redux/states/renderer/annotation.ts
@@ -11,7 +11,7 @@ import { IPQueueState } from "readium-desktop/utils/redux-reducers/pqueue.reduce
 import { INoteCreator } from "../creator";
 // import { IReadiumAnnotation } from "readium-desktop/common/readium/annotation/annotationModel.type";
 import { IColor } from "@r2-navigator-js/electron/common/highlight";
-import { TDrawType } from "./note";
+import type { TDrawType } from "readium-desktop/common/type/note.type";
 
 
 // export const annotationDrawType: TDrawType[] = [
