+++ src/common/redux/states/renderer/note.ts
@@ -12,6 +12,9 @@ import { MiniLocatorExtended } from "../locatorInitialState";
 import { INoteCreator } from "../creator";
 import { IReadiumAnnotation, ISelector } from "readium-desktop/common/readium/annotation/annotationModel.type";
 import type { IPdfTextAnnotationTarget } from "./pdfAnnotation";
+import type { TDrawType } from "readium-desktop/common/type/note.type";
+
+export type { TDrawType } from "readium-desktop/common/type/note.type";
 
 // DO NOT REMOVE THIS COMMENT BLOCK (USED FOR TRANSLATOR KEYS DETECTION DURING CODE SCANNING)
 // __("reader.notes.colors.red")
@@ -64,7 +67,6 @@ export type TNoteColorSet = "pink" | "orange" | "yellow" | "green" | "blue" | "p
 export const NOTE_DEFAULT_COLOR: TNoteColorSet = "yellow";
 export const NOTE_DEFAULT_COLOR_OBJ: IColor = hexToRgb(noteColorSetToColorCode[NOTE_DEFAULT_COLOR]);
 
-export type TDrawType = "solid_background" | "underline" | "strikethrough" | "outline" | "bookmark";
 export const noteDrawType: TDrawType[] = [
     "solid_background",
     "underline",
