+++ src/common/type/note.type.ts
@@ -1,0 +1,8 @@
+// ==LICENSE-BEGIN==
+// Copyright 2017 European Digital Reading Lab. All rights reserved.
+// Licensed to the Readium Foundation under one or more contributor license agreements.
+// Use of this source code is governed by a BSD-style license
+// that can be found in the LICENSE file exposed on Github (readium) in the project repository.
+// ==LICENSE-END==
+
+export type TDrawType = "solid_background" | "underline" | "strikethrough" | "outline" | "bookmark";
