+++ src/renderer/reader/pdf/pdfAnnotationPanel.ts
@@ -5,16 +5,16 @@
 // that can be found in the LICENSE file exposed on Github (readium) in the project repository.
 // ==LICENSE-END==
 
-import type { IColor } from "@r2-navigator-js/electron/common/highlight";
 import { clone } from "ramda";
 
 import { EDrawType } from "readium-desktop/common/redux/states/renderer/note";
 import type { INoteState, TDrawType } from "readium-desktop/common/redux/states/renderer/note";
 import type {
+    IColor,
     TPdfAnnotationNavigationTarget,
     TPdfAnnotationRectTransport,
     TPdfAnnotationSelectionTarget,
-} from "readium-desktop/renderer/reader/pdf/common/pdfReader.type";
+} from "readium-desktop/renderer/reader/pdf/common/pdfAnnotation.type";
 
 export interface IAnnotationPanelSaveChanges {
     color: IColor;
